✨ Put &shy; back where they were after removing&applying suggestions
When we check text, we send the checker a copy that has all soft
hyphens removed, so we won't get botched suggestions (seeing partial
words).

Previously, we would simply strip all soft hyphens from those Editor
paragraphs where we applied any suggestions, so indices would fit when
we got them back from the checker.

This change first builds a map from word-without-shy to word-with-shy
for all words in the document; then for any paragraph where we apply
suggestions, we re-add soft hyphens using that map.

Of course, if we suggest a previously unseen word, that suggestion
won't have any &shy;. We could, instead of the map, use an actual
hyphenator for sme, but now we at least don't mess up words that we
didn't apply suggestions to.